+++ CHANGELOG.md
@@ -1,8 +1,23 @@
-## 0.1.0
+## v0.2.0
+#### INFO
+* Complete rewrite. New parameter-names
+* Existing deployments can't be directly upgraded to this version
+
+#### IMPROVEMENTS
+* Added example provisioning with ansible
+* All relevant variables from provider is included; most are optional
+* Made pinning of hostname optional
+* Better outputs
+* Added automatic testing (basic)
+
+## v0.1.0
+#### IMPROVEMENTS
 * Refactored for Terraform >= 0.12.16
 
-## 0.0.2
+## v0.0.2
+#### BUGS
 * Fixed dns-settings (global)
 
-## 0.0.1
+## v0.0.1
+#### INFO
 * Initial Release